webhook_sg/requirements.txt
flask
gunicorn
tencentcloud-sdk-python
